api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Danny Wilson <dwlsn@pm.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
pyyaml = "^6.0.2"
fastapi = {extras = ["standard"], version = "^0.116.1"}
uvicorn = {extras = ["standard"], version = "^0.35.0"}
fabric = "^3.2.2"
gunicorn = "^23.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
